Use sin/cos instead of sinf/cosf (accuracy)

+++ code/qcommon/q_math.c
@@ -650,10 +650,10 @@ void RotatePointAroundVector( vec3_t dst, const vec3_t dir, const vec3_t point,
 	zrot[0][0] = zrot[1][1] = zrot[2][2] = 1.0F;
 
 	rad = DEG2RAD( degrees );
-	zrot[0][0] = cosf( rad );
-	zrot[0][1] = sinf( rad );
-	zrot[1][0] = -sinf( rad );
-	zrot[1][1] = cosf( rad );
+	zrot[0][0] = cos( rad );
+	zrot[0][1] = sin( rad );
+	zrot[1][0] = -sin( rad );
+	zrot[1][1] = cos( rad );
 
 	Matrix3x3Multiply( m, zrot, tmpmat );
 	Matrix3x3Multiply( tmpmat, im, rot );
@@ -768,14 +768,14 @@ void AnglesToAxis( const vec3_t angles, vec3_t axis[3] ) {
 	// static to help MS compiler fp bugs
 
 	angle = angles[ YAW ] * ( M_PI * 2 / 360 );
-	sy = sinf( angle );
-	cy = cosf( angle );
+	sy = sin( angle );
+	cy = cos( angle );
 	angle = angles[ PITCH ] * ( M_PI * 2 / 360 );
-	sp = sinf( angle );
-	cp = cosf( angle );
+	sp = sin( angle );
+	cp = cos( angle );
 	angle = angles[ ROLL ] * ( M_PI * 2 / 360 );
-	sr = sinf( angle );
-	cr = cosf( angle );
+	sr = sin( angle );
+	cr = cos( angle );
 
 	axis[ 0 ][ 0 ] = cp*cy;
 	axis[ 0 ][ 1 ] = cp*sy;
@@ -903,7 +903,7 @@ the msvc acos doesn't always return a value between -PI and PI:
 
 int i;
 i = 1065353246;
-acosf(*(float*) &i) == -1.#IND0
+acos(*(float*) &i) == -1.#IND0
 
 	This should go in q_math but it is too late to add new traps
 	to game and ui
@@ -912,7 +912,7 @@ acosf(*(float*) &i) == -1.#IND0
 float Q_acos(float c) {
 	float angle;
 
-	angle = acosf(c);
+	angle = acos(c);
 
 	if (angle > M_PI) {
 		return (float)M_PI;
@@ -2051,14 +2051,14 @@ void AngleVectors( const vec3_t angles, vec3_t forward, vec3_t right, vec3_t up 
 	// static to help MS compiler fp bugs
 
 	angle = angles[ YAW ] * ( M_PI * 2 / 360 );
-	sy = sinf( angle );
-	cy = cosf( angle );
+	sy = sin( angle );
+	cy = cos( angle );
 	angle = angles[ PITCH ] * ( M_PI * 2 / 360 );
-	sp = sinf( angle );
-	cp = cosf( angle );
+	sp = sin( angle );
+	cp = cos( angle );
 	angle = angles[ ROLL ] * ( M_PI * 2 / 360 );
-	sr = sinf( angle );
-	cr = cosf( angle );
+	sr = sin( angle );
+	cr = cos( angle );
 
 	if( forward )
 	{
@@ -2087,11 +2087,11 @@ void AngleVectorsLeft( const vec3_t angles, vec3_t forward, vec3_t left, vec3_t 
 	// static to help MS compiler fp bugs
 
 	angle = angles[ YAW ] * ( M_PI * 2 / 360 );
-	sy = sinf( angle );
-	cy = cosf( angle );
+	sy = sin( angle );
+	cy = cos( angle );
 	angle = angles[ PITCH ] * ( M_PI * 2 / 360 );
-	sp = sinf( angle );
-	cp = cosf( angle );
+	sp = sin( angle );
+	cp = cos( angle );
 
 	if( forward )
 	{
@@ -2103,8 +2103,8 @@ void AngleVectorsLeft( const vec3_t angles, vec3_t forward, vec3_t left, vec3_t 
 	if( left || up )
 	{
 		angle = angles[ ROLL ] * ( M_PI * 2 / 360 );
-		sr = sinf( angle );
-		cr = cosf( angle );
+		sr = sin( angle );
+		cr = cos( angle );
 
 		if( left )
 		{
@@ -2717,8 +2717,8 @@ void MatrixSetupXRotation(matrix_t m, vec_t degrees)
 	vec_t a = DEG2RAD(degrees);
 
 	m[ 0] = 1;      m[ 4] = 0;              m[ 8] = 0;              m[12] = 0;
-	m[ 1] = 0;      m[ 5] = cosf(a);         m[ 9] =-sinf(a);         m[13] = 0;
-	m[ 2] = 0;      m[ 6] = sinf(a);         m[10] = cosf(a);         m[14] = 0;
+	m[ 1] = 0;      m[ 5] = cos(a);         m[ 9] =-sin(a);         m[13] = 0;
+	m[ 2] = 0;      m[ 6] = sin(a);         m[10] = cos(a);         m[14] = 0;
 	m[ 3] = 0;      m[ 7] = 0;              m[11] = 0;              m[15] = 1;
 }
 
@@ -2726,9 +2726,9 @@ void MatrixSetupYRotation(matrix_t m, vec_t degrees)
 {
 	vec_t a = DEG2RAD(degrees);
 
-	m[ 0] = cosf(a);         m[ 4] = 0;      m[ 8] = sinf(a);         m[12] = 0;
+	m[ 0] = cos(a);         m[ 4] = 0;      m[ 8] = sin(a);         m[12] = 0;
 	m[ 1] = 0;              m[ 5] = 1;      m[ 9] = 0;              m[13] = 0;
-	m[ 2] =-sinf(a);         m[ 6] = 0;      m[10] = cosf(a);         m[14] = 0;
+	m[ 2] =-sin(a);         m[ 6] = 0;      m[10] = cos(a);         m[14] = 0;
 	m[ 3] = 0;              m[ 7] = 0;      m[11] = 0;              m[15] = 1;
 }
 
@@ -2736,8 +2736,8 @@ void MatrixSetupZRotation(matrix_t m, vec_t degrees)
 {
 	vec_t a = DEG2RAD(degrees);
 
-	m[ 0] = cosf(a);         m[ 4] =-sinf(a);         m[ 8] = 0;      m[12] = 0;
-	m[ 1] = sinf(a);         m[ 5] = cosf(a);         m[ 9] = 0;      m[13] = 0;
+	m[ 0] = cos(a);         m[ 4] =-sin(a);         m[ 8] = 0;      m[12] = 0;
+	m[ 1] = sin(a);         m[ 5] = cos(a);         m[ 9] = 0;      m[13] = 0;
 	m[ 2] = 0;              m[ 6] = 0;              m[10] = 1;      m[14] = 0;
 	m[ 3] = 0;              m[ 7] = 0;              m[11] = 0;      m[15] = 1;
 }
@@ -2939,8 +2939,8 @@ void MatrixToAngles(const matrix_t m, vec3_t angles)
 		sp = -1.0;
 	}
 
-	theta = -asinf(sp);
-	cp = cosf(theta);
+	theta = -asin(sp);
+	cp = cos(theta);
 
 	if(cp > 8192 * FLT_EPSILON)
 	{
@@ -2958,8 +2958,8 @@ void MatrixToAngles(const matrix_t m, vec3_t angles)
 	double          a;
 	double          ca;
 
-	a = asinf(-m[2]);
-	ca = cosf(a);
+	a = asin(-m[2]);
+	ca = cos(a);
 
 	if(fabs(ca) > 0.005)		// Gimbal lock?
 	{
@@ -2983,14 +2983,14 @@ void MatrixFromAngles(matrix_t m, vec_t pitch, vec_t yaw, vec_t roll)
 	static float    sr, sp, sy, cr, cp, cy;
 
     // static to help MS compiler fp bugs
-	sp = sinf(DEG2RAD(pitch));
-	cp = cosf(DEG2RAD(pitch));
+	sp = sin(DEG2RAD(pitch));
+	cp = cos(DEG2RAD(pitch));
 
-	sy = sinf(DEG2RAD(yaw));
-	cy = cosf(DEG2RAD(yaw));
+	sy = sin(DEG2RAD(yaw));
+	cy = cos(DEG2RAD(yaw));
 
-	sr = sinf(DEG2RAD(roll));
-	cr = cosf(DEG2RAD(roll));
+	sr = sin(DEG2RAD(roll));
+	cr = cos(DEG2RAD(roll));
 
 	m[ 0] = cp*cy;  m[ 4] = (sr*sp*cy+cr*-sy);      m[ 8] = (cr*sp*cy+-sr*-sy);     m[12] = 0;
 	m[ 1] = cp*sy;  m[ 5] = (sr*sp*sy+cr*cy);       m[ 9] = (cr*sp*sy+-sr*cy);      m[13] = 0;
@@ -3663,14 +3663,14 @@ void QuatFromAngles(quat_t q, vec_t pitch, vec_t yaw, vec_t roll)
 	static float    sr, sp, sy, cr, cp, cy;
 
 	// static to help MS compiler fp bugs
-	sp = sinf(DEG2RAD(pitch));
-	cp = cosf(DEG2RAD(pitch));
+	sp = sin(DEG2RAD(pitch));
+	cp = cos(DEG2RAD(pitch));
 
-	sy = sinf(DEG2RAD(yaw));
-	cy = cosf(DEG2RAD(yaw));
+	sy = sin(DEG2RAD(yaw));
+	cy = cos(DEG2RAD(yaw));
 
-	sr = sinf(DEG2RAD(roll));
-	cr = cosf(DEG2RAD(roll));
+	sr = sin(DEG2RAD(roll));
+	cr = cos(DEG2RAD(roll));
 
 	q[0] = sr * cp * cy - cr * sp * sy;	// x
 	q[1] = cr * sp * cy + sr * cp * sy;	// y
@@ -3818,19 +3818,19 @@ void QuatToRotAngle( const quat_t q, vec_t *angle ) {
 void QuatToRotAngleAxis( const quat_t q, vec_t *angle, vec3_t axis ) {
 
 	*angle = atan( VectorLength(q)/q[3] );
-	axis[0] = q[0] / sinf(*angle);
-	axis[1] = q[1] / sinf(*angle);
-	axis[2] = q[2] / sinf(*angle);
+	axis[0] = q[0] / sin(*angle);
+	axis[1] = q[1] / sin(*angle);
+	axis[2] = q[2] / sin(*angle);
 
 	*angle *= 360.0f / M_PI;
 }
 
 void QuatFromRotAngleAxis( quat_t q, vec_t angle, const vec3_t axis ) {
 
-	q[0] = axis[0] * sinf( angle*M_PI/360 );
-	q[1] = axis[1] * sinf( angle*M_PI/360 );
-	q[2] = axis[2] * sinf( angle*M_PI/360 );
-	q[3] = cosf( angle*M_PI/360 );
+	q[0] = axis[0] * sin( angle*M_PI/360 );
+	q[1] = axis[1] * sin( angle*M_PI/360 );
+	q[2] = axis[2] * sin( angle*M_PI/360 );
+	q[3] = cos( angle*M_PI/360 );
 }
 
 void QuatToAngles( const quat_t q, vec3_t angles )
@@ -3842,7 +3842,7 @@ void QuatToAngles( const quat_t q, vec3_t angles )
 	q2[ 2 ] = q[ 2 ] * q[ 2 ];
 	q2[ 3 ] = q[ 3 ] * q[ 3 ];
 
-	angles[ PITCH ] = RAD2DEG( asinf( -2 * ( q[ 2 ] * q[ 0 ] - q[ 3 ] * q[ 1 ] ) ) );
+	angles[ PITCH ] = RAD2DEG( asin( -2 * ( q[ 2 ] * q[ 0 ] - q[ 3 ] * q[ 1 ] ) ) );
 	angles[ YAW ] = RAD2DEG( atan2( 2 * ( q[ 2 ] * q[ 3 ] + q[ 0 ] * q[ 1 ] ), ( q2[ 2 ] - q2[ 3 ] - q2[ 0 ] + q2[ 1 ] ) ) );
 	angles[ ROLL ] = RAD2DEG( atan2( 2 * ( q[ 3 ] * q[ 0 ] + q[ 2 ] * q[ 1 ] ), ( -q2[ 2 ] - q2[ 3 ] + q2[ 0 ] + q2[ 1 ] ) ) );
 }
@@ -3869,7 +3869,7 @@ void QuatToAngles( const quat_t q, vec3_t angles )
 	}
 	else
 	{
-		angles[ PITCH ] = asinf( 2.f*( SingularityTest ) ) * RAD_TO_DEG;
+		angles[ PITCH ] = asin( 2.f*( SingularityTest ) ) * RAD_TO_DEG;
 		angles[ YAW ] = atan2( YawY, YawX ) * RAD_TO_DEG;
 		angles[ ROLL ] = atan2( -2.f*( q[W]*q[X] + q[Y]*q[Z] ), ( 1.f - 2.f*( Square( q[X] ) + Square( q[Y] ) ) ) ) * RAD_TO_DEG;
 	}
@@ -4006,8 +4006,8 @@ void QuatSlerp(const quat_t from, const quat_t to, float frac, quat_t out)
 		sinom = 1.0f / sqrtf(sinSqr);
 		omega = atan2f(sinSqr * sinom, absCosom);
 
-		scale0 = sinf((1.0f - frac) * omega) * sinom;
-		scale1 = sinf(frac * omega) * sinom;
+		scale0 = sin((1.0f - frac) * omega) * sinom;
+		scale1 = sin(frac * omega) * sinom;
 	}
 	else
 	{
@@ -4054,8 +4054,8 @@ void MatrixToEulerAngles
 		sp = -1.0;
 	}
 
-	theta = -asinf( sp );
-	cp = cosf( theta );
+	theta = -asin( sp );
+	cp = cos( theta );
 
 	if( cp > 8192 * FLT_EPSILON )
 	{
